+++ moco-core/src/main/java/com/github/dreamhead/moco/HttpRequestExtractor.java
@@ -3,9 +3,9 @@ package com.github.dreamhead.moco;
 import com.google.common.base.Optional;
 
 public abstract class HttpRequestExtractor<T> implements RequestExtractor<T> {
-    protected abstract Optional<T> doExtract(final HttpRequest request);
+    protected abstract Optional<T> doExtract(HttpRequest request);
 
-    public Optional<T> extract(final Request request) {
+    public final Optional<T> extract(final Request request) {
         if (HttpRequest.class.isInstance(request)) {
             return doExtract(HttpRequest.class.cast(request));
         }