+++ moco-core/src/main/java/com/github/dreamhead/moco/dumper/InternalMessageContent.java
@@ -9,12 +9,12 @@ import java.nio.charset.Charset;
 import static com.github.dreamhead.moco.model.MessageContent.content;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public class InternalMessageContent {
+public final class InternalMessageContent {
     private byte[] content;
     private Charset charset;
 
     public InternalMessageContent(@JsonProperty("content") final byte[] content,
-                                  @JsonProperty("charset") Charset charset) {
+                                  @JsonProperty("charset") final Charset charset) {
         this.content = content;
         this.charset = charset;
     }